+++ Assets/Plugins/Juice/Scripts/Runtime/Core/IViewOptions.cs
@@ -2,7 +2,6 @@
 {
 	public interface IViewOptions
 	{
-		Transition InTransition { get; }
-		Transition OutTransition { get; }
+
 	}
 }